Push input/LED and level/state mapping into device
This change pushes the device-oriented mappings of `Input` to `LED` and
`StatusLevel` to `LEDState` down into the `Device` method
`set_input_status_level`. This improves the level of encapsulation and
simplifies interacting with the device. It is also preparation for
implementing a flashing alert status level, where the device will
encapsulate the management of the animation.