Factor out `LEDStaticState` enum
Prior to this change, we were using `LEDState` to hold both the
configured, potentially animated state of a light on the device *and*
the lower level current state of the light, without animation. This
change removes the need to match the animated states (and panic) in
places where those values should not be possible.